Control.Monad.Trans.Unlift

Description

See overview in the README.md

Trans

class (MonadTransControl t, Forall (Identical t)) => MonadTransUnlift t

A monad transformer which can be unlifted, obeying the monad morphism laws.

Since 0.1.0

newtype Unlift t

A function which can move an action down the monad transformer stack, by providing any necessary environment to the action.

Note that, if ImpredicativeTypes worked reliably, this type wouldn't be necessary, and askUnlift would simply include a more generalized type.

Since 0.1.0

Constructors

Unlift

Fields

unlift :: forall a n. Monad n => t n a -> n a

askUnlift :: forall t m. (MonadTransUnlift t, Monad m) => t m (Unlift t)

Get the Unlift action for the current transformer layer.

Since 0.1.0

askRun :: (MonadTransUnlift t, Monad (t m), Monad m) => t m (t m a -> m a)

A simplified version of askUnlift which addresses the common case where polymorphism isn't necessary.

Since 0.1.0

Base

class (MonadBaseControl b m, Forall (IdenticalBase m)) => MonadBaseUnlift b m | m -> b

A monad transformer stack which can be unlifted, obeying the monad morphism laws.

Since 0.1.0

newtype UnliftBase b m

Similar to Unlift, but instead of moving one layer down the stack, moves the action to the base monad.

Since 0.1.0

Constructors

UnliftBase

Fields

unliftBase :: forall a. m a -> b a

askUnliftBase :: forall b m. MonadBaseUnlift b m => m (UnliftBase b m)

Get the UnliftBase action for the current transformer stack.

Since 0.1.0

askRunBase :: MonadBaseUnlift b m => m (m a -> b a)

A simplified version of askUnliftBase which addresses the common case where polymorphism isn't necessary.

Since 0.1.0

Reexports

class MonadTrans t where

The class of monad transformers. Instances should satisfy the following laws, which state that lift is a monad transformation:

lift . return = return

lift (m >>= f) = lift m >>= (lift . f)

Minimal complete definition

lift

Methods

lift :: Monad m => m a -> t m a

Lift a computation from the argument monad to the constructed monad.

class (Applicative b, Applicative m, Monad b, Monad m) => MonadBase b m | m -> b where

Minimal complete definition

liftBase

Methods

liftBase :: b α -> m α

Lift a computation from the base monad

class MonadTrans t => MonadTransControl t where

Minimal complete definition

liftWith, restoreT

Associated Types

type StT (t :: (* -> *) -> * -> *) a :: *

Monadic state of t.

For clarity, because haddock does not display associated types, below are the elaborated StT definitions provided by this library:

StT IdentityT    a ~ a
StT MaybeT       a ~ Maybe a
StT (ErrorT e)   a ~ Error e => Either e a
StT (ExceptT e)  a ~ Either e a
StT ListT        a ~ [a]
StT (ReaderT r)  a ~ a
StT (StateT s)   a ~ (a, s)
StT (WriterT w)  a ~ Monoid w => (a, w)
StT (RWST r w s) a ~ Monoid w => (a, s, w)

Methods

liftWith :: Monad m => (Run t -> m a) -> t m a

liftWith is similar to lift in that it lifts a computation from the argument monad to the constructed monad.

Instances should satisfy similar laws as the MonadTrans laws:

liftWith . const . return = return

liftWith (const (m >>= f)) = liftWith (const m) >>= liftWith . const . f

The difference with lift is that before lifting the m computation liftWith captures the state of t. It then provides the m computation with a Run function that allows running t n computations in n (for all n) on the captured state.

restoreT :: Monad m => m (StT t a) -> t m a

Construct a t computation from the monadic state of t that is returned from a Run function.

Instances should satisfy:

liftWith (\run -> run t) >>= restoreT . return = t

Example type signatures:

restoreT :: Monad m             => m a            -> IdentityT m a
restoreT :: Monad m             => m (Maybe a)    -> MaybeT m a
restoreT :: (Monad m, Error e)  => m (Either e a) -> ErrorT e m a
restoreT :: Monad m             => m (Either e a) -> ExceptT e m a
restoreT :: Monad m             => m [a]          -> ListT m a
restoreT :: Monad m             => m a            -> ReaderT r m a
restoreT :: Monad m             => m (a, s)       -> StateT s m a
restoreT :: (Monad m, Monoid w) => m (a, w)       -> WriterT w m a
restoreT :: (Monad m, Monoid w) => m (a, s, w)    -> RWST r w s m a

class MonadBase b m => MonadBaseControl b m | m -> b where

Minimal complete definition

liftBaseWith, restoreM

Associated Types

type StM (m :: * -> *) a :: *

Monadic state of m.

For all non-transformer monads, StM m a ~ a:

StM IO         a ~ a
StM Maybe      a ~ a
StM (Either e) a ~ a
StM []         a ~ a
StM ((->) r)   a ~ a
StM Identity   a ~ a
StM STM        a ~ a
StM (ST s)     a ~ a

All transformer monads' StM depends on both the monadic state of the transformer (given by its StT from MonadTransControl), as well as its inner monad's monadic state, given by its StM from MonadBaseControl:

StM (IdentityT  m) a ~ StM m a
StM (MaybeT     m) a ~ StM m (Maybe a)
StM (ErrorT e   m) a ~ Error e => StM m (Either e a)
StM (ExceptT e  m) a ~ StM m (Either e a)
StM (ListT      m) a ~ StM m [a]
StM (ReaderT r  m) a ~ StM m a
StM (StateT s   m) a ~ StM m (a, s)
StM (WriterT w  m) a ~ Monoid w => StM m (a, w)
StM (RWST r w s m) a ~ Monoid w => StM m (a, s, w)

Methods

liftBaseWith :: (RunInBase m b -> b a) -> m a

liftBaseWith is similar to liftIO and liftBase in that it lifts a base computation to the constructed monad.

Instances should satisfy similar laws as the MonadIO and MonadBase laws:

liftBaseWith . const . return = return

liftBaseWith (const (m >>= f)) = liftBaseWith (const m) >>= liftBaseWith . const . f

The difference with liftBase is that before lifting the base computation liftBaseWith captures the state of m. It then provides the base computation with a RunInBase function that allows running m computations in the base monad on the captured state.

restoreM :: StM m a -> m a

Construct a m computation from the monadic state of m that is returned from a RunInBase function.

Instances should satisfy:

liftBaseWith (\runInBase -> runInBase m) >>= restoreM = m